EL-209: update conf and still in wip for test with mock
