+++ .babelrc
@@ -1,4 +1,3 @@
 {
-    "presets": ["@babel/preset-typescript"]
-  }
-  
+  "presets": ["@babel/preset-typescript"]
+}
