--- jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    preset: 'ts-jest',
-    transform: {
-      '^.+\\.(ts|tsx)?$': 'ts-jest',
-      "^.+\\.(js|jsx)$": "babel-jest",
-    }
-  };
-
